+++ iowmongotools/__init__.py
@@ -1,6 +1,6 @@
 """ Main module """
 __author__ = "Denis Ashcheulov"
-__version__ = "0.5.0"
+__version__ = "0.5.1"
 __status__ = "Pre-Alpha"
 
 import logging